lab6/agent1/requirements.txt
google-adk
python-dotenv
pytz
tzdata
fastapi
uvicorn
sqlalchemy